Project intro
Project step 1
Date: January 28, 2019
Introduction:
My name is Lien Zhang. This is an introduction to my project ideas. I will be walking through my project ideas. This idea will be consist of what type it is, the intension that it will be designed for and a brief description of how to do that. 
I will choose to create an application to create on-the-fly invoices that could be used by small business owners to generate professional invoices from their phones. It would have data entry points that were connected with a professional invoice template and as you enter data it would automatically start forming a .pdf that could be accessed at a later time.
The project data will consist of all information that to build a robust database. Data relating to the customers information will be categorized into different type, region, website, functionality, etc. Then the invoice will be created by the invoice generator from the stored data. 

The user interface would be a graphical one, so that a regular person can figure out quite easily and be able to input information and manager their business. This app will be highly customizable which client can go in and update or change their data and setting to better secure and manage.   
This first step in my plan is make charts of the databases and how they will work together. Then to learn about the graphical interfaces and how to connect that to all of the databases. Then go several rounds of testing to ensure that they work together. I hoping that by the end of this project I will have a fully functional program.